@@ -4,12 +4,12 @@ build-backend = "setuptools.build_meta"
 
 [tool.black]
 line-length = 120
-target-version = ['py38']
+target-version = ['py310']
 extend-exclude = "angrmanagement/vendor/"
 
 [tool.ruff]
 line-length = 120
-target-version = 'py38'
+target-version = 'py310'
 exclude = ["angrmanagement/vendor/"]
 
 [tool.ruff.lint]
