@@ -39,7 +39,7 @@ install_requires =
     pyzmq>=17.1
     packaging
 
-python_requires = >=3.8
+python_requires = >=3.10
 include_package_data = True
 
 [options.entry_points]
