@@ -4,7 +4,7 @@ import contextlib
 import logging
 import os
 import re
-from typing import TYPE_CHECKING, Any, Callable
+from typing import TYPE_CHECKING, Any
 
 import tomlkit
 import tomlkit.exceptions
@@ -18,6 +18,7 @@ from .config_entry import ConfigurationEntry as CE
 
 if TYPE_CHECKING:
     import enum
+    from collections.abc import Callable
 
 _l = logging.getLogger(__name__)
 color_re = re.compile("[0-9a-fA-F]+")
