@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING, Generator
+from typing import TYPE_CHECKING
 
 from angr import KnowledgeBase, Project
 from angr.analyses.reaching_definitions.call_trace import CallTrace
@@ -28,6 +28,8 @@ except ImportError:
     argument_resolver = None
 
 if TYPE_CHECKING:
+    from collections.abc import Generator
+
     from angr.analyses.reaching_definitions import ReachingDefinitionsAnalysis
     from angr.knowledge_plugins.key_definitions.atoms import Atom
 
