@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import logging
 import os
-from typing import TYPE_CHECKING, Sequence
+from typing import TYPE_CHECKING
 
 from angrmanagement.data.breakpoint import BreakpointType
 from angrmanagement.data.trace import BintraceTrace
@@ -10,6 +10,8 @@ from angrmanagement.data.trace import BintraceTrace
 from .debugger import Debugger
 
 if TYPE_CHECKING:
+    from collections.abc import Sequence
+
     from angr import SimState
     from angr.knowledge_plugins import Function
 
