@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import os
-from typing import TYPE_CHECKING, Sequence
+from typing import TYPE_CHECKING
 
 from PySide6.QtCore import QSize, Qt
 from PySide6.QtGui import QIcon
@@ -34,6 +34,8 @@ from angrmanagement.data.analysis_options import (
 )
 
 if TYPE_CHECKING:
+    from collections.abc import Sequence
+
     from angrmanagement.ui.workspace import Workspace
 
 
