@@ -147,7 +147,7 @@ class LoadBinary(QDialog):
             dep_list.addItem(dep_item)
 
         if partial_ld.main_object is not None:
-            if isinstance(partial_ld.main_object, (cle.MetaELF, cle.PE, cle.MachO, cle.CGC)):
+            if isinstance(partial_ld.main_object, cle.MetaELF | cle.PE | cle.MachO | cle.CGC):
                 self._base_addr = partial_ld.main_object.mapped_base
                 self._entry_addr = partial_ld.main_object.entry
             else:
