@@ -1,13 +1,15 @@
 from __future__ import annotations
 
 from functools import partial
-from typing import TYPE_CHECKING, Callable
+from typing import TYPE_CHECKING
 
 from angrmanagement.config import Conf
 
 from .menu import Menu, MenuEntry, MenuSeparator
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
+
     from angrmanagement.ui.views.disassembly_view import DisassemblyView
 
 
