@@ -1,11 +1,14 @@
 from __future__ import annotations
 
 from functools import partial
-from typing import Callable
+from typing import TYPE_CHECKING
 
 from angrmanagement.config import Conf
 
 from .menu import Menu, MenuEntry, MenuSeparator
+
+if TYPE_CHECKING:
+    from collections.abc import Callable
 
 
 class DisasmLabelContextMenu(Menu):
