@@ -190,7 +190,7 @@ class DataDepView(InstanceView):
     def _convert_node(
         self, node: BaseDepNode, converted: dict[BaseDepNode, QDataDepGraphBlock]
     ) -> QDataDepGraphBlock | None:
-        if isinstance(node, (MemDepNode, RegDepNode)):
+        if isinstance(node, MemDepNode | RegDepNode):
             cs_instr = self._instructions.get(node.ins_addr, None)
             instr = cs_instr.insn if cs_instr else None
         else:
