@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Sequence
+from typing import TYPE_CHECKING
 
 from PySide6.QtCore import QSize
 from PySide6.QtGui import QAction
@@ -10,6 +10,8 @@ from angrmanagement.data.object_container import ObjectContainer
 from angrmanagement.ui.icons import icon
 
 if TYPE_CHECKING:
+    from collections.abc import Sequence
+
     import angr
     import PySide6.QtGui
 
