@@ -1,10 +1,12 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Callable
+from typing import TYPE_CHECKING
 
 from PySide6.QtWidgets import QLineEdit
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
+
     from angrmanagement.data.instance import Instance
 
 
