@@ -79,7 +79,9 @@ class QPhiVariable(QCachedGraphicsItem):
         painter.drawText(x, self._config.disasm_font_ascent, "\u0278(")
         x += self._config.disasm_font_width * 2
 
-        for i, (subvar_ident, ident_width) in enumerate(zip(self._subvar_idents, self._subvar_ident_widths)):
+        for i, (subvar_ident, ident_width) in enumerate(
+            zip(self._subvar_idents, self._subvar_ident_widths, strict=False)
+        ):
             painter.setPen(Qt.darkGreen)
             painter.drawText(x, self._config.disasm_font_ascent, subvar_ident)
             x += ident_width
